optimistic updates
In this section, swapped more functionality over to using the API. We now use the database to:

remove Todos and Goals
toggle the state of a Todos
save a new Todo or Goal
What's important is that for the removing and toggling, we're doing these actions optimistically. So we're assuming the change will succeed correctly on the server, so we update the UI immediately, and then only roll back to the original state if the API returns an error. Doing optimistic updates is better because it provides a more realistic and dynamic experience to the user.